+++ extras/install_chef_server.sh
@@ -1,18 +1,20 @@
 #!/bin/bash -x
 
-mkdir -p /vagrant/packages
-cd /vagrant/packages
-echo downloading chef-server debian package.  this may take some time...
-wget -nc --quiet https://opscode-omnitruck-release.s3.amazonaws.com/ubuntu/12.04/x86_64/chef-server_11.0.6-1.ubuntu.12.04_amd64.deb
-dpkg -i chef-server*.deb
-chef-server-ctl reconfigure
+if [ ! -f /home/vagrant/.chef/knife.db ]
+then
+  echo "Preparing to install chef client..."
+  mkdir -p /vagrant/packages
+  cd /vagrant/packages
+  echo downloading chef-server debian package.  this may take some time...
+  wget -nc --quiet https://opscode-omnitruck-release.s3.amazonaws.com/ubuntu/12.04/x86_64/chef-server_11.0.6-1.ubuntu.12.04_amd64.deb
+  dpkg -i chef-server*.deb
+  chef-server-ctl reconfigure
 
-
-mkdir -p /home/vagrant/.chef
-sudo cp /etc/chef-server/admin.pem /home/vagrant/.chef/
-sudo cp /etc/chef-server/chef-validator.pem /home/vagrant/.chef/
-mkdir -p /vagrant/.chef
-cp /home/vagrant/.chef/* /vagrant/.chef/
+  mkdir -p /home/vagrant/.chef
+  sudo cp /etc/chef-server/admin.pem /home/vagrant/.chef/
+  sudo cp /etc/chef-server/chef-validator.pem /home/vagrant/.chef/
+  mkdir -p /vagrant/.chef
+  cp /home/vagrant/.chef/* /vagrant/.chef/
 
 cat<<KNIFE > /home/vagrant/.chef/knife.rb
 log_level                :info
@@ -26,5 +28,9 @@ cache_type               'BasicFile'
 cache_options( :path => '/home/vagrant/.chef/checksums' )
 KNIFE
 
-chown vagrant. /vagrant/.chef/*
-chown vagrant. /home/vagrant/.chef/*
+  chown vagrant. /vagrant/.chef/*
+  chown vagrant. /home/vagrant/.chef/*
+
+else 
+  echo "chef server already installed!"
+fi
